feat: comments modal

+++ hone/src/components/Entry.tsx
@@ -1,5 +1,5 @@
 import { FC, useEffect, useState, useRef, Dispatch } from "react";
-import { Entry as EntryInterface, Image } from "../globals";
+import { Entry as EntryInterface, Image, Comment as CommentInterface } from "../globals";
 import "../styles/entry.css";
 import { storage } from '../firebase';
 import { ref, getDownloadURL, uploadBytes, getStorage, deleteObject } from "firebase/storage";
@@ -7,10 +7,14 @@ import {
   Modal,
   ModalOverlay,
   ModalContent,
+  ModalHeader,
   ModalBody,
+  ModalFooter,
   ModalCloseButton,
   useDisclosure
 } from '@chakra-ui/react'
+import Comment from "./Comment";
+import { useAuth } from "../hooks/useAuth";
 
 type Props = {
   entry: EntryInterface;
@@ -26,15 +30,15 @@ const Entry: FC<Props> = ({ entry, setEntries, isSameUser, isCommentsOn }) => {
   const [newEntryDescription, setNewEntryDescription] = useState<string>(entry.description);
   const [newEntryImage, setNewEntryImage] = useState<File>();
   const [dateCreatedString, setDateCreated] = useState<string>("");
-  const [comments, setComments] = useState<Comment>();
-  
-  
+  const [comments, setComments] = useState<CommentInterface[]>([]); //use array
+  const [newComment, setNewComment] = useState<string>("");
+  const { user, isLoggedIn } = useAuth();
+
   const inputImage = useRef(null);
 
   const { isOpen, onOpen, onClose } = useDisclosure()
-
-  //Kiarosh
   const { isOpen: isCommentsOpen, onOpen: onCommentsOpen, onClose: onCommentsClose } = useDisclosure();
+  const { isOpen: isDeleteOpen, onOpen: onDeleteOpen, onClose: onDeleteClose } = useDisclosure();
 
   useEffect(() => {
     async function fetchImage() {
@@ -50,7 +54,12 @@ const Entry: FC<Props> = ({ entry, setEntries, isSameUser, isCommentsOn }) => {
     setDateCreated(setDateString);
 
     if (isCommentsOn) {
-      // fetch /commments/entries/:entryId
+      async function fetchComments() {
+        const fetchComment = await fetch(`${process.env.API_URL}/comments/entries/${entry?.id}`);
+        const comments = await fetchComment.json();
+        setComments(comments);
+      }
+      fetchComments();
     }
 
   }, [])
@@ -128,13 +137,16 @@ const Entry: FC<Props> = ({ entry, setEntries, isSameUser, isCommentsOn }) => {
       method: "DELETE",
     })
 
-    await fetch(`${process.env.API_URL}/images/${image?.id}`, {
-      method: "DELETE",
-    })
+    if (image) {
+      await fetch(`${process.env.API_URL}/images/${image?.id}`, {
+        method: "DELETE",
+      });
 
-    const storage = getStorage();
-    const deleteRef = ref(storage, image?.filePath);
-    deleteObject(deleteRef);
+      const storage = getStorage();
+      const deleteRef = ref(storage, image?.filePath);
+      deleteObject(deleteRef);
+    }
+
   }
 
   function handleChange(event: React.ChangeEvent<HTMLInputElement>) { // Upload image
@@ -143,8 +155,29 @@ const Entry: FC<Props> = ({ entry, setEntries, isSameUser, isCommentsOn }) => {
     setNewEntryImage(imageToUpload);
   }
 
+  async function handleCommentOnClick() {
+    const body = { description: newComment, user_id: user?.id, entry_id: entry.id }
+    if (newComment !== "") {
+      const fetchNewComment = await fetch(`${process.env.API_URL}/comments`, {
+        method: "POST",
+        headers: {
+          "Content-Type": "application/json",
+        },
+        body: JSON.stringify(body),
+      })
+
+      const newComment = await fetchNewComment.json();
+      setComments((prev) => {
+        const newArray = [...prev];
+        newArray.push(newComment);
+        return newArray;
+      });
+    }
+    setNewComment("");
+  }
+
   return (
-    <>
+    <section>
       <div className="entry-container">
         {image ? <img src={image?.url} className="entry-img" onClick={onOpen} /> : null}
         <div className="entry-date-description-container">
@@ -152,9 +185,8 @@ const Entry: FC<Props> = ({ entry, setEntries, isSameUser, isCommentsOn }) => {
             <p className="entry-date">{dateCreatedString}</p>
             {/* <RelativeTime date={entry.created_date} /><hr /> */}
             {(isSameUser && !isEditable) ? <button className="edit-entry-btn" onClick={handleEditOnClick}><span className="material-symbols-outlined">edit</span></button> : null}
-            {isSameUser ? <button className="edit-entry-btn" onClick={handleDeleteOnClick}><span className="material-symbols-outlined">delete</span></button> : null}
-            {/* //**********************Kiarosh */}
-            {(isCommentsOn) ? <button className="view-comments-btn" onClick={onCommentsOpen}><span className="material-symbols-outlined">comment</span></button>: null}
+            {isSameUser ? <button className="edit-entry-btn" onClick={onDeleteOpen}><span className="material-symbols-outlined">delete</span></button> : null}
+            {(isCommentsOn) ? <button className="edit-entry-btn" onClick={onCommentsOpen}><span className="material-symbols-outlined">comment</span></button> : null}
           </div>
           <hr />
           {isEditable ? <textarea id="editable-entry-description" className="editable-entry-description" value={newEntryDescription} onChange={(e) => setNewEntryDescription(e.target.value)} autoFocus /> : <p className="entry-p"> {entryDescription}</p>}
@@ -165,17 +197,28 @@ const Entry: FC<Props> = ({ entry, setEntries, isSameUser, isCommentsOn }) => {
         </div>
       </div>
 
-  <Modal isOpen={isCommentsOpen} onClose={onCommentsClose}>
-    <ModalOverlay />
-    <ModalContent>
-      <ModalCloseButton />
-        <ModalBody>
-          <h2>Test</h2>
-      </ModalBody>
-    </ModalContent>
-  </Modal>
+      <Modal isOpen={isCommentsOpen} onClose={onCommentsClose} autoFocus={false} returnFocusOnClose={false}>
+        <ModalOverlay />
+        <ModalContent className="comments-modal">
+          <ModalCloseButton className="modal-close-btn" />
+          <ModalHeader>Comments</ModalHeader>
+          <ModalBody className="comments-modal-body">
+            {comments?.map((comment) => (
+              <Comment key={comment.id} comment={comment} setComments={setComments} />
+            ))}
+          </ModalBody>
+          <ModalFooter>
+            {isLoggedIn ? <div className="comment-input-submit-container">
+              <input className="comment-input" type="text" value={newComment} onChange={(e) => setNewComment(e.target.value)} autoFocus placeholder="Add a comment..." />
+              <button onClick={handleCommentOnClick} className="comment-submit-btn">Submit</button>
+            </div> :
+              <input type="text" placeholder="Log in to add a comment!" disabled className="disabled-comment-input" />
+            }
+          </ModalFooter>
+        </ModalContent>
+      </Modal>
 
-      <Modal isOpen={isOpen} onClose={onClose}>
+      <Modal isOpen={isOpen} onClose={onClose} autoFocus={false} returnFocusOnClose={false}>
         <ModalOverlay />
         <ModalContent maxH="90vh" maxW="90vw" color="transparent" bg="transparent" alignItems={"center"} boxShadow={"none"}>
           <ModalCloseButton margin="0" boxShadow={"none"} bg="white" outline={"transparent"} />
@@ -184,7 +227,24 @@ const Entry: FC<Props> = ({ entry, setEntries, isSameUser, isCommentsOn }) => {
           </ModalBody>
         </ModalContent>
       </Modal>
-    </>
+
+      <Modal isOpen={isDeleteOpen} onClose={onDeleteClose}>
+        <ModalOverlay />
+        <ModalContent>
+          <ModalHeader marginTop="0.5em">Are you sure you want to delete <br />this entry?</ModalHeader>
+          <ModalCloseButton />
+          <ModalBody>
+            This action cannot be undone!
+          </ModalBody>
+          <ModalFooter>
+            <div className="btn-container">
+              <button className="modal-btn cancel-btn" onClick={onDeleteClose}>Cancel</button>
+              <button className="modal-btn" id="delete-project-btn" onClick={handleDeleteOnClick}>Delete</button>
+            </div>
+          </ModalFooter>
+        </ModalContent>
+      </Modal>
+    </section>
   )
 }
